add datastore

+++ src/downloader.rs
@@ -50,7 +50,7 @@ impl Downloader {
                 Message::Interested =>    state.peer_interested = true,
                 Message::NotInterested => state.peer_interested = false,
                 Message::Bitfield { bits } => {
-                    if bits.len() < info.num_pieces() {
+                    if (bits.len() as u64) < info.num_pieces() {
                         println!("{}", Error::new_str(&format!("bitfield has less bits {} than pieces {}",
                                                            bits.len(), info.num_pieces())));
                     }
